requirements.txt
playwright>=1.43.0,<2.0.0
pandas>=2.2.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
openai>=1.30.0,<2.0.0
